# api/internal/httpapi/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Splatmaker API
  version: 0.2.0
  description: Core-first API module with async orchestration and idempotent job submission.
servers:
  - url: /
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthResponse:
      type: object
      required: [status, mode]
      properties:
        status:
          type: string
        mode:
          type: string
    JobStatus:
      type: string
      enum: [new, queued, in_progress, done, failed, cancelled]
    JobSummary:
      type: object
      required: [job_id, status, progress_percent, created_at, updated_at]
      properties:
        job_id:
          type: string
        status:
          $ref: '#/components/schemas/JobStatus'
        progress_percent:
          type: integer
          minimum: 0
          maximum: 100
        current_step:
          type: string
          nullable: true
        idempotency_key:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    JobDetails:
      type: object
      required: [summary, attempt, simulate_failure, output_files]
      properties:
        summary:
          $ref: '#/components/schemas/JobSummary'
        attempt:
          type: integer
          minimum: 1
        source_ref:
          type: string
          nullable: true
        simulate_failure:
          type: boolean
        error_message:
          type: string
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        last_heartbeat_at:
          type: string
          format: date-time
          nullable: true
        output_files:
          type: array
          items:
            $ref: '#/components/schemas/OutputFileRef'
    OutputFileRef:
      type: object
      required: [key, file_name]
      properties:
        key:
          type: string
        file_name:
          type: string
        size_bytes:
          type: integer
          nullable: true
    ListJobsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/JobSummary'
    SubmitJobRequest:
      type: object
      required: [idempotency_key]
      properties:
        idempotency_key:
          type: string
          minLength: 1
        name:
          type: string
          nullable: true
        source_ref:
          type: string
          nullable: true
        simulate_failure:
          type: boolean
          default: false
    SubmitJobResponse:
      type: object
      required: [created, job]
      properties:
        created:
          type: boolean
        job:
          $ref: '#/components/schemas/JobDetails'
    JobMutationResponse:
      type: object
      required: [job]
      properties:
        job:
          $ref: '#/components/schemas/JobDetails'
    ResultFileURL:
      type: object
      required: [key, file_name, url, expires_at]
      properties:
        key:
          type: string
        file_name:
          type: string
        url:
          type: string
        expires_at:
          type: string
          format: date-time
    JobResultURLsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ResultFileURL'
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [user_id, email]
      properties:
        user_id:
          type: string
        email:
          type: string
    LoginResponse:
      type: object
      required: [access_token, token_type, user]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        user:
          $ref: '#/components/schemas/AuthUser'
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ReconstructionSoftwareName:
      type: string
      enum: [glomap, colmap]
    ReconstructionMatchingMethod:
      type: string
      enum: [sequential, exhaustive, spatial, vocab_tree, transitive]
    TrainingModel:
      type: string
      enum: [splatfacto]
    Training3DIsp:
      type: string
      enum: [none]
    PostProcessingCropMode:
      type: string
      enum: [environment]
    CoordinateSystem:
      type: string
      enum: [rhyu]
    BackgroundRemovalModel:
      type: string
      enum: [u2net]
    ObjectRemovalAction:
      type: string
      enum: [erase]
    PipelineSettings:
      type: object
      required: [videoProcessing, reconstruction, training, postProcessing, sphericalCamera, segmentation]
      properties:
        videoProcessing:
          type: object
          required: [maxNumImages, videoStartTime, videoStopTime, filterBlurryImages]
          properties:
            maxNumImages: { type: integer }
            videoStartTime: { type: integer }
            videoStopTime: { type: integer, description: "-1 means unset" }
            filterBlurryImages: { type: boolean }
        reconstruction:
          type: object
          required: [enable, softwareName, enableEnhancedFeatureExtraction, matchingMethod, enableFlHeuristic, flHeuristicValue, enableFlMetric, flMetricValue, posePriors]
          properties:
            enable: { type: boolean }
            softwareName: { $ref: '#/components/schemas/ReconstructionSoftwareName' }
            enableEnhancedFeatureExtraction: { type: boolean }
            matchingMethod: { $ref: '#/components/schemas/ReconstructionMatchingMethod' }
            enableFlHeuristic: { type: boolean }
            flHeuristicValue: { type: number }
            enableFlMetric: { type: boolean }
            flMetricValue: { type: number }
            posePriors:
              type: object
              required: [usePosePriorColmapModelFiles, usePosePriorTransformJson]
              properties:
                usePosePriorColmapModelFiles: { type: boolean }
                usePosePriorTransformJson:
                  type: object
                  required: [enable, sourceCoordinateName, poseIsWorldToCam]
                  properties:
                    enable: { type: boolean }
                    sourceCoordinateName: { type: string }
                    poseIsWorldToCam: { type: boolean }
        training:
          type: object
          required: [enable, maxSteps, model, 3dIsp, preserveSceneScale, enableDepthLoss]
          properties:
            enable: { type: boolean }
            maxSteps: { type: integer }
            model: { $ref: '#/components/schemas/TrainingModel' }
            3dIsp: { $ref: '#/components/schemas/Training3DIsp' }
            preserveSceneScale: { type: boolean }
            enableDepthLoss: { type: boolean }
        postProcessing:
          type: object
          required: [cropOutputBounds, cropMode, cleanSplat, enableSpz, enableSog, enableUsdz, enableVideoExport, plyCoords, spzCoords, sogCoords, usdzCoords]
          properties:
            cropOutputBounds: { type: boolean }
            cropMode: { $ref: '#/components/schemas/PostProcessingCropMode' }
            cleanSplat: { type: boolean }
            enableSpz: { type: boolean }
            enableSog: { type: boolean }
            enableUsdz: { type: boolean }
            enableVideoExport: { type: boolean }
            plyCoords: { $ref: '#/components/schemas/CoordinateSystem' }
            spzCoords: { $ref: '#/components/schemas/CoordinateSystem' }
            sogCoords: { $ref: '#/components/schemas/CoordinateSystem' }
            usdzCoords: { $ref: '#/components/schemas/CoordinateSystem' }
        sphericalCamera:
          type: object
          required: [enable, cubeFacesToRemove, optimizeSequentialFrameOrder]
          properties:
            enable: { type: boolean }
            cubeFacesToRemove: { type: string }
            optimizeSequentialFrameOrder: { type: boolean }
        segmentation:
          type: object
          required: [backgroundRemoval, objectRemoval]
          properties:
            backgroundRemoval:
              type: object
              required: [enable, model, maskThreshold]
              properties:
                enable: { type: boolean }
                model: { $ref: '#/components/schemas/BackgroundRemovalModel' }
                maskThreshold: { type: number }
            objectRemoval:
              type: object
              required: [enable, action, objects]
              properties:
                enable: { type: boolean }
                action: { $ref: '#/components/schemas/ObjectRemovalAction' }
                objects: { type: string }
    PipelineSettingsResponse:
      type: object
      required: [settings]
      properties:
        settings:
          $ref: '#/components/schemas/PipelineSettings'
paths:
  /healthz:
    get:
      operationId: healthz
      summary: Health check
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /v1/auth/login:
    post:
      operationId: login
      summary: Username/password login (standalone)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '501':
          description: Not implemented for current runtime mode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/pipeline-settings/standard:
    get:
      operationId: getStandardPipelineSettings
      summary: Get standard pipeline settings (falls back to defaults when absent)
      responses:
        '200':
          description: Standard pipeline settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineSettingsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: putStandardPipelineSettings
      summary: Save standard pipeline settings for current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PipelineSettings'
      responses:
        '200':
          description: Standard pipeline settings saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineSettingsResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/jobs:
    get:
      operationId: listJobs
      summary: List current user jobs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Jobs list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListJobsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: submitJob
      summary: Submit a reconstruction job asynchronously (idempotent by idempotency_key)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitJobRequest'
      responses:
        '202':
          description: Job queued or idempotent replay resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitJobResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/jobs/{job_id}:
    get:
      operationId: getJob
      summary: Get details for a single job
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDetails'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/jobs/{job_id}/cancel:
    post:
      operationId: cancelJob
      summary: Cancel a queued/running job
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobMutationResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Job cannot be cancelled in current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/jobs/{job_id}/retry:
    post:
      operationId: retryJob
      summary: Retry a failed/cancelled job
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job queued for retry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobMutationResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Job cannot be retried in current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/jobs/{job_id}/result-urls:
    get:
      operationId: getJobResultUrls
      summary: Resolve result file URLs for a job
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
        - name: ttl_seconds
          in: query
          schema:
            type: integer
            default: 900
            minimum: 1
      responses:
        '200':
          description: Result URLs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResultURLsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
x-splatmaker-runtime:
  mode: standalone